+++ backend/src/models.rs
@@ -2,7 +2,7 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct User {
-    pub id: i32,
+    pub id: u64,
     pub username: String,
     pub watchlist: Vec<String>,
     pub altitude: bool,
